requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
python-dotenv==1.0.0
boto3==1.21.32
botocore==1.24.32
pandas==2.1.3
Flask==2.0.1
Werkzeug==2.0.3
gunicorn==20.1.0
flask-login==0.5.0
stripe==2.60.0

# Security packages
flask-talisman==1.1.0
cryptography==41.0.7
pyOpenSSL==23.3.0
flask-seasurf==1.1.1
